+++ PyMOPanel/constants.py
