+++ PyMOPanel/helpers.py
