+++ demo_filesystem.py
@@ -1,17 +1,17 @@
 #!/usr/bin/env python3
 from time import sleep
 from sys import argv
-from PyMOPanel import MatrixOrbital
+from PyMOPanel import MatrixOrbital as Panel, filesystem as fs
 
 def main(port):
-    myPanel = MatrixOrbital(port=port)
+    myPanel = Panel(port=port)
 
     # dump complete filesystem to a file
-    myPanel.dumpCompleteFilesystem('filesystem.data')
+    fs.dumpAll(myPanel, 'filesystem.data')
 
-    print("filesystem free space: {} bytes".format(myPanel.getFilesystemFreeSpaceInBytes()))
+    print("filesystem free space: {} bytes".format(fs.getFreeSpaceInBytes(myPanel)))
 
-    filesystemContent = myPanel.getFilesystemDirectory()
+    filesystemContent = fs.getDirectory(myPanel)
     print("filesystem content: {}".format(str(filesystemContent)))
     
     print("Downloading all files:")
@@ -21,7 +21,7 @@ def main(port):
             continue
         outputFilename = '{}_{}.data'.format(fileType.name, str(fileId))
         print("Writting {} with size {}.".format(outputFilename, fileSize))
-        myPanel.downloadFile(fileType, fileId, outputFilename)
+        fs.download(myPanel, fileType, fileId, outputFilename)
 
 if __name__ == '__main__':
     port = argv[1] if len(argv) == 2 else '/dev/ttyUSB0'
